requirements.txt
torch>=2.4.0
transformers>=4.51.0
datasets>=3.0.0
peft>=0.14.0
trl>=0.15.0
bitsandbytes>=0.45.0
accelerate>=1.3.0
vllm>=0.8.0
openai>=1.60.0
anthropic>=0.42.0
aiohttp>=3.10.0
click>=8.1.0
pydantic>=2.10.0
scikit-learn>=1.5.0
pandas>=2.2.0
tqdm>=4.67.0
python-dotenv>=1.0.0
huggingface-hub>=0.28.0
safetensors>=0.5.0
